# backend/YAESandBox.API.DTOs/BlockDtos.cs
using YAESandBox.Core.Block;
using YAESandBox.Depend;

// For BlockStatusCode

namespace YAESandBox.API.DTOs;

/// <summary>
/// 用于 API 响应，表示单个 Block 的详细信息（不包含 WorldState）。
/// </summary>
public record BlockDetailDto
{
    /// <summary>
    /// Block 的唯一标识符。
    /// </summary>
    public required string BlockId { get; init; } = null!;

    /// <summary>
    /// 父 Block 的 ID。如果为根节点，则为 null。
    /// </summary>
    public string? ParentBlockId { get; init; }

    /// <summary>
    /// Block 当前的状态码 (例如 Idle, Loading, ResolvingConflict, Error)。
    /// </summary>
    public BlockStatusCode? StatusCode { get; init; }

    /// <summary>
    /// Block 的主要文本内容 (例如 AI 生成的文本、配置等)。
    /// </summary>
    public string? BlockContent { get; init; }

    /// <summary>
    /// 与 Block 相关的元数据字典 (键值对均为字符串)。
    /// </summary>
    public Dictionary<string, string>? Metadata { get; init; }

    /// <summary>
    /// 该 Block 的直接子 Block 的 ID 列表。
    /// </summary>
    public List<string>? ChildrenInfo { get; init; }
    // 注意：WsInput, WsPostAI, WsPostUser, WsTemp 不应直接通过 API 暴露
}

/// <summary>
/// 用于标识 Block 中可能过时的字段。
/// </summary>
public enum BlockDataFields
{
    ParentBlockId,
    BlockContent,
    Metadata,
    ChildrenInfo,
    WorldState,
    GameState
}

/// <summary>
/// 用于通过 PATCH 请求部分更新 Block 的内容和元数据。
/// 任何设置为 null 的属性表示不修改该部分。
/// </summary>
public record UpdateBlockDetailsDto
{
    /// <summary>
    /// (可选) 要设置的新的 Block 内容。
    /// 如果为 null，则不修改 BlockContent。
    /// </summary>
    public string? Content { get; init; }

    /// <summary>
    /// (可选) 要更新或移除的元数据键值对。
    /// - Key: 要操作的元数据键。
    /// - Value:
    ///   - 如果为非 null 字符串: 添加或更新该键的值。
    ///   - 如果为 null: 从元数据中移除该键。
    /// 如果整个字典为 null，则不修改 Metadata。
    /// </summary>
    public Dictionary<string, string?>? MetadataUpdates { get; init; }
}