# frontend/src/api-client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "YAESandBox API (Public)",
    "description": "供前端和外部使用的公开 API 文档。",
    "version": "v1"
  },
  "paths": {
...
    "/api/manage/blocks/{blockId}": {
      "delete": {
        "tags": [
          "BlockManagement"
        ],
        "summary": "手动删除一个指定的 Block。",
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "description": "要删除的 Block ID。",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recursive",
            "in": "query",
            "description": "是否递归删除子 Block。默认递归删除，非递归可能导致奇奇怪怪的问题？",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "force",
            "in": "query",
            "description": "是否强制删除，无视状态。",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作已成功执行。"
          },
          "400": {
            "description": "不允许删除根节点或请求无效（例如 Block 有子节点但未指定 recursive=true）。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "未找到具有指定 ID 的 Block。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Block的当前状态不允许删除。现在只允许删除Idle或Error，说实在的有点意义不明，也许以后设计成什么都可以删除。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "执行操作时发生内部服务器错误。"
          }
        }
      }
    },
...
    "/api/Blocks/{blockId}": {
      "get": {
        "tags": [
          "Blocks"
        ],
        "summary": "获取指定 ID 的单个 Block 的详细信息（不包含 WorldState）。",
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "description": "要查询的 Block 的唯一 ID。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回 Block 详细信息。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BlockDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "未找到具有指定 ID 的 Block。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Blocks"
        ],
        "summary": "部分更新指定 Block 的内容和/或元数据。\r\n此操作仅在 Block 处于 Idle 状态时被允许。",
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "description": "要更新的 Block 的 ID。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "包含要更新的字段（Content, MetadataUpdates）的请求体。\r\n            省略的字段或值为 null 的字段将不会被修改（MetadataUpdates 中值为 null 表示移除该键）。",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlockDetailsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlockDetailsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlockDetailsDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "更新成功。"
          },
          "400": {
            "description": "请求体无效或未提供任何更新。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "未找到具有指定 ID 的 Block。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Block 不处于 Idle 状态，无法修改。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "更新时发生内部服务器错误。"
          }
        }
      }
    },
    "/api/Blocks/topology": {
      "get": {
        "tags": [
          "Blocks"
        ],
        "summary": "获取整个 Block 树的拓扑结构 (基于 ID 的嵌套关系)。\r\n返回一个表示 Block 树层级结构的 JSON 对象。",
        "parameters": [
          {
            "name": "blockId",
            "in": "query",
            "description": "目标根节点的ID，如果为空则返回整个父节点的ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回 JSON 格式的拓扑结构。\r\n            形如：{ \"id\": \"__WORLD__\", \"children\": [{ \"id\": \"child1\", \"children\": [] },{ \"id\": \"child2\", \"children\": [] }] }",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonBlockNode"
                }
              }
            }
          },
          "500": {
            "description": "生成拓扑结构时发生内部服务器错误。"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlockDetailDto": {
        "required": [
          "blockId"
        ],
        "type": "object",
        "properties": {
          "blockId": {
            "type": "string",
            "description": "Block 的唯一标识符。",
            "nullable": true
          },
          "parentBlockId": {
            "type": "string",
            "description": "父 Block 的 ID。如果为根节点，则为 null。",
            "nullable": true
          },
          "statusCode": {
            "$ref": "#/components/schemas/BlockStatusCode"
          },
          "blockContent": {
            "type": "string",
            "description": "Block 的主要文本内容 (例如 AI 生成的文本、配置等)。",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "与 Block 相关的元数据字典 (键值对均为字符串)。",
            "nullable": true
          },
          "childrenInfo": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "该 Block 的直接子 Block 的 ID 列表。",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "用于 API 响应，表示单个 Block 的详细信息（不包含 WorldState）。"
      },
      "BlockStatusCode": {
        "enum": [
          "Loading",
          "Idle",
          "ResolvingConflict",
          "Error",
          "NotFound",
          "Deleted"
        ],
        "type": "string",
        "description": "表示 Block 的不同状态。"
      },
...
      "JsonBlockNode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JsonBlockNode"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "UpdateBlockDetailsDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "(可选) 要设置的新的 Block 内容。\r\n如果为 null，则不修改 BlockContent。",
            "nullable": true
          },
          "metadataUpdates": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "description": "(可选) 要更新或移除的元数据键值对。\r\n- Key: 要操作的元数据键。\r\n- Value:\r\n  - 如果为非 null 字符串: 添加或更新该键的值。\r\n  - 如果为 null: 从元数据中移除该键。\r\n如果整个字典为 null，则不修改 Metadata。",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "用于通过 PATCH 请求部分更新 Block 的内容和元数据。\r\n任何设置为 null 的属性表示不修改该部分。"
      },
...
    }
  }
}